Generate JWT based on claim.
