+++ mesos-client/build.gradle
@@ -1,6 +1,7 @@
 dependencies {
     testCompile project(':test-utils')
     implementation "com.lightbend.akka:akka-stream-alpakka-simple-codecs_$scalaVersion:1.0-M2"
+    implementation "com.pauldijou:jwt-json4s-native_$scalaVersion:2.1.0"
 
     // Transitive dependencies
     compile "org.apache.mesos:mesos:$mesosVersion"
